Expose necessary attributes to do streaming replication, bump version
Previously, any configuration options pertinent to
using Pgpool with a streaming replication cluster
were not exposed anywhere, leaving one to fend for
themselves via role/environment overrides. This
problem is solved for here by exposing those
options and others as default attributes, as well
as relevant documentation on the options that
should not be left at their defaults.